Create badge service; refactor some controller actions
