+++ app/controllers/admin/badges_controller.rb
@@ -1,4 +1,39 @@
 # frozen_string_literal: true
 
 class Admin::BadgesController < Admin::BaseController
+  before_action :set_badge, only: %i[show edit update destroy]
+
+  def index
+    @badges = Badge.all
+  end
+
+  def show; end
+
+  def new
+    @badge = Badge.new
+  end
+
+  def edit; end
+
+  def create
+
+  end
+
+  def update
+
+  end
+
+  def destroy
+
+  end
+
+  private
+
+  def set_badge
+    @badge = Badge.find(params[:id])
+  end
+
+  def badge_params
+    params.require(:badge).permit(:name, :image, :description)
+  end
 end
