+++ app/controllers/user_badges_controller.rb
@@ -1,4 +1,8 @@
 # frozen_string_literal: true
 
 class UserBadgesController < ApplicationController
+  def index
+    @user_badges = current_user.badges
+    @badges = Badge.all
+  end
 end
