+++ app/services/test_passages/create_badge_service.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+module TestPassages
+  class CreateBadgeService
+
+
+    def initialize(test_passage)
+      @test_passage = test_passage
+
+    end
+
+    def call
+
+    end
+
+    private
+
+    def method
+
+    end
+  end
+end
